Ajout de la méthode validationFinal et optimisation de certaine partie du code

+++ projet-equipe14/src/FormationContinue.java
@@ -31,7 +31,7 @@ public class FormationContinue {
         return heuresTransferees;
     }
 
-    public void setHeuresTransferees(long heuresTransferees) {
+    public void setHeuresTransferees(int heuresTransferees) {
         this.heuresTransferees = heuresTransferees;
     }
 }
